cups: Fix missing installation of sysvinit scripts
Configuration with --without-systemd causes the SYSTEMD_DIR to have a
value of "no"

cups configure test for, if SYSTEMD_DIR is empty to decide if the init
scripts need to be installed. A value of "no" results in that no init
scripts is installed.

Removing --without-systemd and just keep it empty result in an empty
SYSTEMD_DIR variable and the sysvinit scripts is installed as expected.
Though the cups configure script could be improved to work consistent
with "--without-*" options.

Signed-off-by: Claus Stovgaard <claus.stovgaard@gmail.com>
Signed-off-by: Ross Burton <ross.burton@arm.com>