lighttpd: rationalise module packaging
There's no need to explicitly set FILES, we can just pass prepend=True
when splitting.  This means FILES:${PN} can just add /www.

The extra_depends='' simply means the modules have no RDEPENDS, but this
is pointless as the modules are useless without lighttpd itself. Remove
that so the modules RDEPEND on lighttpd.

Move the module depends/recommends to the bottom of the recipe where
they belong.  Also recommend instead of depending on dirlisting, this is
not a mandatory module and there are many good reasons to not enable them.

Signed-off-by: Ross Burton <ross.burton@arm.com>