libunwind: upgrade to 1.8.0
Change the SRC_URI because the new tarball isn't available at
nongnu.org.

Rewrite and submit upstream a PPC patch. PPC32 on musl is the only
configuration which needs libatomic_ops so limit the DEPENDS
appropriately.

Remove manual linkage to libucontext with musl, backport a patch from
upstream to do this instead.

Drop 0001-src-Gtrace-remove-unguarded-print-calls.patch, merged.
fix status

Always enable the documentation: latex2man is only needed to rebuild the
documentation, but the manpages are shipped in the tarball. This needs a
small patch.

Signed-off-by: Ross Burton <ross.burton@arm.com>