libexif: improve ptest packaging
The NLS tests don't actually test anything useful and were not being ran
anyway, so don't install them.

Install the test binaries using libtool so that we install the actual
binaries instead of libtool wrapper scripts.  This means that if we also
tell the Makefile to use /bin/sh there's no need to RDEPEND on bash.

Signed-off-by: Ross Burton <ross.burton@arm.com>