meta: uboot-config: Fix broken menuconfig for "fitImage" kernel
kernel-fitimage inherited uboot-config, which resulted in failure to
run "menuconfig" on a system that has both multiple u-boot configs and
wants a fitImage kernel.

Fix the issue by splitting up uboot-config into conf/uboot.conf that
only contains the "harmless" variable settings, and require that in a
new uboot-config class which contains the Python code to do the U-boot
specific settings. Require only conf/uboot.conf in kernel-fitimage so
that it has what it needs but doesn't break.

Fixes: f9e834e31788 "meta: introduce KCONFIG_CONFIG_ENABLE_MENUCONFIG"
Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>